fix:merged-fix

+++ backend/src/main/java/com/goorm/goormfriends/api/dto/request/UpdateDirectoryRequest.java
@@ -6,7 +6,7 @@ import lombok.ToString;
 @Getter
 @ToString
 public class UpdateDirectoryRequest {
-    private int userId;
-    private int directoryId;
+    private Long userId;
+    private Long directoryId;
     private String newDirectoryTitle;
 }
